Trait subspace_proof_of_space::Table

pub trait Table: Sized + Send + Sync + 'static {
    type Generator: TableGenerator<Self>;
    type Quality<'a>: Quality
       where Self: 'a;

    const TABLE_TYPE: PosTableType;

    // Required methods
    fn generate(seed: &PosSeed) -> Self;
    fn find_quality(&self, challenge_index: u32) -> Option<Self::Quality<'_>>;
    fn is_proof_valid(
        seed: &PosSeed,
        challenge_index: u32,
        proof: &PosProof
    ) -> Option<PosQualityBytes>;

    // Provided methods
    fn generate_parallel(seed: &PosSeed) -> Self { ... }
    fn generator() -> Self::Generator { ... }
}

Proof of space kind

Required Associated Types

type Generator: TableGenerator<Self>

Instance that can be used to generate tables with better performance

type Quality<'a>: Quality where Self: 'a

Abstraction that represents quality of the solution in the table

Required Associated Constants

const TABLE_TYPE: PosTableType

Proof of space table type

Required Methods

fn generate(seed: &PosSeed) -> Self

Generate new table with 32 bytes seed.

There is also Self::generate_parallel() that can achieve lower latency.

fn find_quality(&self, challenge_index: u32) -> Option<Self::Quality<'_>>

Try to find quality of the proof at challenge_index if proof exists

fn is_proof_valid( seed: &PosSeed, challenge_index: u32, proof: &PosProof ) -> Option<PosQualityBytes>

Check whether proof created earlier is valid and return quality bytes if yes

Provided Methods

fn generate_parallel(seed: &PosSeed) -> Self

Generate new table with 32 bytes seed using parallelism.

This implementation will trade efficiency of CPU and memory usage for lower latency, prefer Self::generate() unless lower latency is critical.

fn generator() -> Self::Generator

Returns a stateful table generator with better performance